fixed handling of slices between form and ruleset
Gave ruleset objects an attribute `string_slice` to hold the input from
the form, in string format, rather than passing to the actual
`index_slice` dictionary which is used to slice the dataframe.

This means that when fields are passed back to the form, they are in the
right format.

New bug tho: slicing doesn't seem to work with individual product names