added scenario support
A bit hacky.  Because of all the to and fro between rulesets and the
form, loading scenario happens in two parts.  First load the rulesets
and make the (empty) form.  Then populate the form (i.e. for all
rulesets).

Also had to create a new attribute of RuleSet class: `func_str` to hold
the string identifier of the function.  Actually works better than
looking up the function repr.